Clean up CMR
